move Header component to its own file and folder
